11. Building the Scheduling System | 108. Step 20 Extracting the Table Availability Logic Into its Own Function
