+++ app/pages/api/restaurant/[slug]/availability.ts
@@ -1,4 +1,4 @@
-import { times } from '@/app/data';
+import { findAvailableTables } from '@/app/services/restaurant/findAvailableTables';
 import { PrismaClient } from '@prisma/client';
 import { NextApiRequest, NextApiResponse } from 'next';
 
@@ -21,42 +21,6 @@ export default async function handler(
     });
   }
 
-  const searchTimes = times.find((t) => {
-    return t.time === time;
-  })?.searchTimes;
-
-  if (!searchTimes) {
-    return res.status(400).json({
-      errorMessage: 'Invalid data provided',
-    });
-  }
-
-  const bookings = await prisma.booking.findMany({
-    where: {
-      booking_time: {
-        gte: new Date(`${day}T${searchTimes[0]}`),
-        lte: new Date(`${day}T${searchTimes[searchTimes.length - 1]}`),
-      },
-    },
-    select: {
-      number_of_people: true,
-      booking_time: true,
-      tables: true,
-    },
-  });
-
-  const bookingTablesObj: { [key: string]: { [key: number]: true } } = {};
-
-  bookings.forEach((booking) => {
-    bookingTablesObj[booking.booking_time.toISOString()] =
-      booking.tables.reduce((obj, table) => {
-        return {
-          ...obj,
-          [table.table_id]: true,
-        };
-      }, {});
-  });
-
   const restaurant = await prisma.restaurant.findUnique({
     where: {
       slug,
@@ -74,24 +38,18 @@ export default async function handler(
     });
   }
 
-  const tables = restaurant.tables;
-
-  const searchTimesWithTables = searchTimes.map((searchTime) => {
-    return {
-      date: new Date(`${day}T${searchTime}`),
-      time: searchTime,
-      tables,
-    };
+  const searchTimesWithTables = await findAvailableTables({
+    day,
+    time,
+    res,
+    restaurant,
   });
 
-  searchTimesWithTables.forEach((t) => {
-    t.tables = t.tables.filter((table) => {
-      if (bookingTablesObj[t.date.toISOString()]) {
-        if (bookingTablesObj[t.date.toISOString()][table.id]) return false;
-      }
-      return true;
+  if (!searchTimesWithTables) {
+    return res.status(400).json({
+      errorMessage: 'Invalid data provided',
     });
-  });
+  }
 
   const availabilities = searchTimesWithTables
     .map((t) => {
